requirements.txt
fastmcp==2.*
uvicorn
httpx
python-dotenv
Pillow
playwright